requirements.txt
asgiref==3.8.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
crispy-bootstrap4==2025.6
cryptography==45.0.5
Django==5.2.3
django-allauth==65.10.0
django-ckeditor-5==0.2.18
django-crispy-forms==2.4
django-js-asset==3.1.2
django-taggit==6.1.0
idna==3.10
oauthlib==3.3.1
pillow==11.2.1
pycparser==2.22
PyJWT==2.10.1
requests==2.32.4
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
